Fixed GFlop calculation output

+++ fft1d.C
@@ -42,7 +42,7 @@ struct Main : public CBase_Main {
 
     void done() {
       double time = CkWallTimer() - start;
-      double gflops = 5*N*N*log2(N*N)/(time*1000000000);
+      double gflops = 5*(double)N*N*log2((double)N*N)/(time*1000000000);
       CkPrintf("chares: %d\ncores: %d\nsize: %d\ntime: %f sec\nrate: %f GFlop/s\n",
         numChares, CkNumPes(), N*N, time, gflops);
       fftProxy.writeResults(0);
